Completed code rafactoring

+++ routes/groups.py
@@ -1,6 +1,6 @@
-from globals import app, db, timestamp
+from globals import app, db, timestamp, get_arg_or_400
 from flask_login import login_required, current_user
-from flask import render_template, request, redirect, url_for, flash
+from flask import render_template, request, redirect, url_for, flash, abort
 from forms import NewGroupFrom
 from libs.Group import Group
 from libs.GroupMember import GroupMember
@@ -11,46 +11,51 @@ from libs.Invitation import InvitationType, Invitation
 @login_required
 def group_route():
 
-    def args_error():
-        flash("Invalid request", 'error')
-        return redirect(url_for(request.referrer))
-
     if request.method == 'GET':
-        action = request.args.get('action')
+        action = get_arg_or_400('action')
+
         if not action:
             flash("Something went wrong!", "error")
             return redirect(url_for('group_route', action='my'))
+
         elif action == 'my':
             return render_template('my_groups.html', groups=current_user.get_groups())
-        if action == 'new':
+
+        elif action == 'new':
             new_group_form = NewGroupFrom()
             return render_template('new_group.html', form=new_group_form, groups=current_user.get_groups())
 
-        try:
-            group_id = int(request.args.get('id'))
-        except ValueError:
-            return args_error()
-        group = Group.get(group_id)
+        group_id = get_arg_or_400('id', to_int=True)
+        group = Group.get_or_404(group_id)
         members = group.get_members()
         is_member = current_user in members
         events = group.get_events()
         is_admin = group.admin_id == current_user.id
-        print(f"events = {events}")
         if not is_member:
             is_member = None
-        elif action == 'delete':
+
+        if action == 'delete' and is_admin:
             group.delete()
             return redirect(url_for('group_route', action='my'))
-        elif action == 'edit':
-            # TODO add edit form
+
+        elif action == 'edit' and is_admin:
             edit_group_form = NewGroupFrom(
                 closed=group.closed,
                 name=group.name,
                 sport=group.sport
             )
             return render_template('edit_group.html', form=edit_group_form, group=group)
+
         elif action == 'show':
-            pass
+            return render_template(
+                'group.html',
+                group=group,
+                members=members,
+                is_member=is_member,
+                events=events,
+                is_admin=is_admin
+            )
+
         elif action == 'join':
             if current_user not in members:
                 if group.closed:
@@ -62,6 +67,15 @@ def group_route():
                     db.session.commit()
                     members.append(current_user)
                     is_member = True
+            return render_template(
+                'group.html',
+                group=group,
+                members=members,
+                is_member=is_member,
+                events=events,
+                is_admin=is_admin
+            )
+
         elif action == 'leave':
             if current_user in members:
                 row = GroupMember.query.filter_by(user_id=current_user.id, group_id=group.id).first()
@@ -69,24 +83,23 @@ def group_route():
                 db.session.commit()
                 members.remove(current_user)
                 is_member = None
-        return render_template(
-            'group.html',
-            group=group,
-            members=members,
-            is_member=is_member,
-            events=events,
-            is_admin=is_admin
-        )
+            return render_template(
+                'group.html',
+                group=group,
+                members=members,
+                is_member=is_member,
+                events=events,
+                is_admin=is_admin
+            )
+        else:
+            abort(403)
+
     else:
-        type = request.args.get('type')
-        if type is None:
-            return args_error()
-        elif type == 'edit':
-            try:
-                group_id = int(request.args.get('id'))
-            except ValueError:
-                return args_error()
-            group = Group.get(group_id)
+        type = get_arg_or_400('type')
+
+        if type == 'edit':
+            group_id = get_arg_or_400('id', to_int=True)
+            group = Group.get_or_404(group_id)
 
             edit_group_form = NewGroupFrom(current_group=group)
             if edit_group_form.validate_on_submit():
@@ -96,7 +109,9 @@ def group_route():
                 db.session.add(group)
                 db.session.commit()
                 return redirect(url_for('group_route', action='show', id=group_id))
-            return render_template('edit_group.html', form=edit_group_form, group=group)
+            else:
+                return render_template('edit_group.html', form=edit_group_form, group=group)
+
         elif type == 'new':
             new_group_form = NewGroupFrom()
             if new_group_form.validate_on_submit():
@@ -112,4 +127,8 @@ def group_route():
                 db.session.add(new_row)
                 db.session.commit()
                 return redirect(url_for('group_route', action='show', id=group.id))
-            return render_template('new_group.html', form=new_group_form, groups=current_user.get_groups())
+            else:
+                return render_template('new_group.html', form=new_group_form, groups=current_user.get_groups())
+
+        else:
+            abort(400)
